Add more functionalities to the project
The project now has several buttons:
-One button to draw a new grid, which prompts the
user for the number of squares.
-One button to clear the sketch pad so the current
drawing is erased.
-One button to toggle between light green color
and random color. Either color mode have a
darkening effect implemented, which progressively
darkens the square if already colored.

The project also has a functionality to toggle
the pad on and off. This is achieved with the
key 't'. This is useful so that the user can draw
anywhere in the pad without having to leave a
streak when entering the pad from outside and
getting to where they want to draw.